+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/splunk/DisabledSplunkForwarder.kt
@@ -3,9 +3,11 @@ package com.atlassian.performance.tools.infrastructure.api.splunk
 import com.atlassian.performance.tools.ssh.api.SshConnection
 
 class DisabledSplunkForwarder : SplunkForwarder {
-
-
     override fun run(sshConnection: SshConnection, name: String) {
+        return
+    }
+
+    override fun run(sshConnection: SshConnection, name: String, logsPath: String) {
         return
     }