+++ CHANGELOG.md
@@ -29,6 +29,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ### Added
 - Add the `CPU_UTILISATION` dimension.
 
+### Changed
+- Changes in `SplunkForwarder`. Added explicit parameter with logs path. Added explicit parameters to `AtlassianSplunkForwarder` constructor.
+
 
 ## [2.1.1] - 2018-09-11
 [2.1.1]: https://bitbucket.org/atlassian/infrastructure/branches/compare/release-2.1.1%0Drelease-2.1.0